Commands and base type was separated from each other; even more Cleaned up a bit to be more human-readable
